Executor trait refactor for harvester, mapping, enrichement, jsonl and remap entries Move summary logging methods and pingTwofishes into Utils
